fix(ssg): split bare page render and refresh API docs after rebase
Keep generateHtmlPage's a11y, team, and pageChrome parameters in the
generated reference. Move BarePageData out of render.rs so the file
stays under the 350-line limit.
